update __main__ to call run_server script
