+++ pyproject.toml
@@ -42,10 +42,9 @@ dev = [
 Repository = "https://github.com/simon-mcintosh/imas-mcp-server"
 
 [project.scripts]
-imas-mcp-server = "imas_mcp_server.mcp_imas:run_server"
-run-server = "imas_mcp_server.mcp_imas:run_server"
-index-name = "scripts.index_name:main"
-build-index = "scripts.build_index:main"
+run-server = "scripts.run_server:run_server"
+index-name = "scripts.index_name:index_name"
+build-index = "scripts.build_index:build_index"
 
 [tool.hatch.version]
 source = "vcs"
